refactor: move types to types folder
