+++ src/services/lab-admin/proctorApiService.ts
@@ -1,43 +1,17 @@
+import { ProctorData, ProctorCreateResponse } from "@/types/proctors";
 import { apiService, ApiResponse } from "../apiService";
-
-// Proctor data interface - represents all the form fields
-export interface ProctorData {
-  jobNumber: string;
-  proctorTestNumber: string;
-  materialType: string;
-  dateSampled: string;
-  proctorType: "SPDD" | "MPDD";
-  maxDryDensity: string;
-  correctedDensity: string;
-  labLocation: string;
-  proctorId: string;
-  dateTested: string;
-  oversizePercentage: number;
-  optimumMoisture: number;
-  specificGravity: string;
-}
-
-// API response interfaces
-export interface ProctorCreateResponse {
-  id: string;
-  message: string;
-  proctor: ProctorData;
-}
-
-export interface ProctorListResponse {
-  proctors: ProctorData[];
-  total: number;
-}
 
 // Main ProctorApiService class
 class ProctorApiService {
   private readonly baseEndpoint = "/api/proctors/lab-admin";
 
   // Get all proctors for a job
-  async getAllProctors(jobId: number): Promise<ApiResponse<ProctorData[]>> {
+  async getAllProctors(jobNumber: string): Promise<ApiResponse<ProctorData[]>> {
     try {
-      console.log("Getting all proctors for the job id: ", jobId);
-      return await apiService.get<ProctorData[]>(`/api/proctors/job/${jobId}`);
+      console.log("Getting all proctors for the job id: ", jobNumber);
+      return await apiService.get<ProctorData[]>(
+        `/api/proctors/job/${jobNumber}`,
+      );
     } catch (error) {
       console.error("Error fetching proctors for a job id", error);
