+++ src/types/photos.ts
@@ -1,0 +1,7 @@
+export type Photo = {
+  id: number;
+  title: string;
+  updated: string;
+  src: string;
+  description: string;
+};
