+++ src/types/sitePlan.ts
@@ -1,0 +1,6 @@
+export type SitePlan = {
+  id: number;
+  name: string;
+  src: string;
+  dateCreated: string;
+};
